# perf: speed up method `LayoutPostprocessor._process_special_clusters`