@cau-git 2025-08-19
@mohammedahmed18 Thanks for taking effort to find optimization potential in docling's code base. We will consider optimization PRs where the balance between increase in code complexity and total runtime saved is striking a good balance. Looking at the proposed code change, I can see that the `possible` table candidates are found by assuming _full enclosure_ in the wrapper (full _and_ conjunction), which is not exactly the logic of the original code. The original code would tolerate an overlap of 0.9 which is including nearly-enclosed tables. Hence this would need some additional considerations before we merge.